+++ modules/kubernetes/starboard/main.tf
@@ -69,8 +69,9 @@ resource "helm_release" "trivy" {
   version     = "0.4.12"
   max_history = 3
   values = [templatefile("${path.module}/templates/trivy-values.yaml.tpl", {
-    provider       = var.cloud_provider
-    trivy_role_arn = var.trivy_role_arn
+    provider                        = var.cloud_provider
+    trivy_role_arn                  = var.trivy_role_arn
+    volume_claim_storage_class_name = var.volume_claim_storage_class_name
   })]
 }