+++ modules/aws/eks-core/route53.tf
@@ -1,15 +1,15 @@
 resource "aws_route53_zone" "this" {
-  name = var.dns_zone
-
   for_each = {
-    for s in ["zone"] :
-    s => s
+    for dns in var.dns_zone :
+    dns => dns
   }
+
+  name = each.key
 
   tags = merge(
     local.global_tags,
     {
-      Name = var.dns_zone
+      Name = each.key
     },
   )
 }