+++ modules/kubernetes/cert-manager/variables.tf
@@ -18,14 +18,14 @@ variable "azure_config" {
   description = "Azure specific configuration"
   type = object({
     subscription_id     = string,
-    hosted_zone_name    = string,
+    hosted_zone_names   = list(string),
     resource_group_name = string,
     client_id           = string,
     resource_id         = string,
   })
   default = {
     subscription_id     = "",
-    hosted_zone_name    = "",
+    hosted_zone_names   = [],
     resource_group_name = "",
     client_id           = "",
     resource_id         = "",
@@ -36,12 +36,12 @@ variable "aws_config" {
   description = "AWS specific configuration"
   type = object({
     region         = string,
-    hosted_zone_id = string,
+    hosted_zone_id = map(string)
     role_arn       = string,
   })
   default = {
     region         = "",
-    hosted_zone_id = "",
+    hosted_zone_id = {},
     role_arn       = "",
   }
 }