+++ modules/kubernetes/eks-core/variables.tf
@@ -54,7 +54,7 @@ variable "aad_groups" {
 variable "kubernetes_network_policy_default_deny" {
   description = "If network policies should by default deny cross namespace traffic"
   type        = bool
-  default     = false
+  default     = true
 }
 
 variable "kubernetes_default_limit_range" {