+++ modules/azure/aks-global/velero.tf
@@ -1,12 +1,13 @@
 #tfsec:ignore:azure-storage-queue-services-logging-enabled
 resource "azurerm_storage_account" "velero" {
-  name                     = "strg${var.environment}${var.location_short}${var.name}velero${var.unique_suffix}"
-  resource_group_name      = data.azurerm_resource_group.this.name
-  location                 = data.azurerm_resource_group.this.location
-  account_tier             = "Standard"
-  account_replication_type = "GRS"
-  account_kind             = "StorageV2"
-  min_tls_version          = "TLS1_2"
+  name                            = "strg${var.environment}${var.location_short}${var.name}velero${var.unique_suffix}"
+  resource_group_name             = data.azurerm_resource_group.this.name
+  location                        = data.azurerm_resource_group.this.location
+  account_tier                    = "Standard"
+  account_replication_type        = "GRS"
+  account_kind                    = "StorageV2"
+  min_tls_version                 = "TLS1_2"
+  allow_nested_items_to_be_public = false
 }
 
 resource "azurerm_storage_container" "velero" {