+++ modules/azure/azure-pipelines-agent-vmss/main.tf
@@ -13,7 +13,7 @@ terraform {
 
   required_providers {
     azurerm = {
-      version = "2.99.0"
+      version = "3.1.0"
       source  = "hashicorp/azurerm"
     }
     tls = {