+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Fix
+
+- [#614](https://github.com/XenitAB/terraform-modules/pull/620) Fix broken cluster autoscaler
+
 ## 2022.03.5
 
 ### Changed