+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fix
 
 - [#614](https://github.com/XenitAB/terraform-modules/pull/620) Fix broken cluster autoscaler
+- [#617](https://github.com/XenitAB/terraform-modules/pull/617) Upgrade falco to 0.31.1
 
 ## 2022.03.5