+++ CHANGELOG.md
@@ -9,10 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - [#622](https://github.com/XenitAB/terraform-modules/pull/622) [Breaking] Hardcode prometheus and trivy storage class.
+- [#617](https://github.com/XenitAB/terraform-modules/pull/617) Upgrade falco to 0.31.1
 
 ### Fix
 
-- [#614](https://github.com/XenitAB/terraform-modules/pull/620) Fix broken cluster autoscaler
+- [#620](https://github.com/XenitAB/terraform-modules/pull/620) Fix broken cluster autoscaler
 
 ## 2022.03.5