+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - [#624](https://github.com/XenitAB/terraform-modules/pull/624) Add Promtail for platform logs in Azure.
+- [#630](https://github.com/XenitAB/terraform-modules/pull/630) Add variable for exluding Promtail namespaces
 
 ### Changed